requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0
anthropic>=0.40.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Frontend
streamlit>=1.39.0
streamlit-chat>=0.1.1

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1

# Database & Storage
chromadb>=0.5.0
sqlite-vec>=0.1.0

# Web & API
requests>=2.32.0
httpx>=0.27.0
beautifulsoup4>=4.12.0

# GitHub Integration
PyGithub>=2.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
pydantic>=2.9.0
typing-extensions>=4.12.0
